+++ tests/minimal_project/build.py
@@ -9,8 +9,6 @@ sys.path.insert(0, SRBUILD_DIR)
 # Normal build files would start here:
 import srbuild
 
-# srbuild.logger.G_LOGGER.severity = srbuild.logger.G_LOGGER.VERBOSE
-
 project = srbuild.Project()
 libmath = project.library("math", sources=["factorial.cpp", "fibonacci.cpp"], libs=["stdc++"])
 test = project.executable("test", sources=["test.cpp"], libs=["stdc++", libmath])